Move getfullargspec into compat module
