+++ ddtrace/internal/compat.py
@@ -1,3 +1,4 @@
+from inspect import isgeneratorfunction
 import platform
 import random
 import re
@@ -83,6 +84,27 @@ if PYTHON_VERSION_INFO >= (3, 7):
     pattern_type = re.Pattern
 else:
     pattern_type = re._pattern_type  # type: ignore[misc,attr-defined]
+
+try:
+    from inspect import getargspec as getfullargspec
+
+    def args_provided(f, argspec):
+        return argspec.args or argspec.varargs or argspec.keywords or argspec.defaults or isgeneratorfunction(f)
+
+
+except ImportError:
+    from inspect import getfullargspec  # type: ignore[misc]  # noqa: F401
+
+    def args_provided(f, argspec):
+        return (
+            argspec.args
+            or argspec.varargs
+            or argspec.varkw
+            or argspec.defaults
+            or argspec.kwonlyargs
+            or argspec.kwonlydefaults
+            or isgeneratorfunction(f)
+        )
 
 
 def is_integer(obj):
